made project compilable

+++ migrations/001_Init.py
@@ -45,7 +45,7 @@ def migrate(migrator, database, fake=False, **kwargs):
 
     @migrator.create_model
     class User(pw.Model):
-        id = pw.IntegerField(primary_key=True)
+        id = pw.BigIntegerField(primary_key=True)
         name = pw.CharField(max_length=255)
         username = pw.CharField(max_length=255)
         language = pw.CharField(constraints=[SQL("DEFAULT 'en'")], default='en', max_length=255)
